Modulize lowering logic
Split the native_lowering.py monolith into a module with files handling the lowering procedure of step types.
