+++ SymbolicDSGE/monte_carlo/native_lowering/__init__.py
@@ -1,0 +1,9 @@
+"""Native Monte Carlo pipeline lowering.
+
+The package keeps the public lowering boundary compact while separating the
+native ABI contracts for simulations, filters, regressions, and diagnostics.
+"""
+
+from .core import FloatInputBinding, LoweredMCRun, lower_native_run
+
+__all__ = ["FloatInputBinding", "LoweredMCRun", "lower_native_run"]
